Simplify minirent consumption
This change simplifies the integration with this library. It refactors
`#if` guards to automatically include `<dirent.h>` on non-WIN32
platforms.

Therefore, usages of this library can be changed from this:

```c
 #ifdef _WIN32
 #  define MINIRENT_IMPLEMENTATION
 #  include "minirent.h"
 #else
 #  include <dirent.h>
 #endif
```

to this:

```c
 #define MINIRENT_IMPLEMENTATION
 #include "minirent.h"
```

No fuss, no wuss. This should be backwards compatible. README and CI
were updated to reflect it.